LayoutTests/http/tests/loading/resourceLoadStatistics/user-interaction-in-cross-origin-sub-frame-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
frame "testFrame" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "testFrame" - didCommitLoadForFrame
frame "testFrame" - didFinishDocumentLoadForFrame
frame "testFrame" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "testFrame" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
Tests that user interaction is always logged for the top document/frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.location.origin is topFrameOrigin
PASS testRunner.isStatisticsHasHadUserInteraction(topFrameOrigin) is false
PASS testRunner.isStatisticsHasHadUserInteraction(subFrameOrigin) is false
PASS testRunner.isStatisticsHasHadUserInteraction(topFrameOrigin) is true
PASS testRunner.isStatisticsHasHadUserInteraction(subFrameOrigin) is false
PASS successfullyParsed is true

TEST COMPLETE
